+++ providers/keycloak_oidc.go
@@ -8,7 +8,10 @@ import (
 	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/sessions"
 )
 
-const keycloakOIDCProviderName = "Keycloak OIDC"
+const (
+	keycloakOIDCProviderName = "Keycloak OIDC"
+	keycloakOIDCDefaultScope = "openid email profile"
+)
 
 // KeycloakOIDCProvider creates a Keycloak provider based on OIDCProvider
 type KeycloakOIDCProvider struct {
@@ -18,7 +21,8 @@ type KeycloakOIDCProvider struct {
 // NewKeycloakOIDCProvider makes a KeycloakOIDCProvider using the ProviderData
 func NewKeycloakOIDCProvider(p *ProviderData, opts options.KeycloakOptions) *KeycloakOIDCProvider {
 	p.setProviderDefaults(providerDefaults{
-		name: keycloakOIDCProviderName,
+		name:  keycloakOIDCProviderName,
+		scope: keycloakOIDCDefaultScope,
 	})
 
 	provider := &KeycloakOIDCProvider{